Upgrade to anchor 0.29.0. Optionally update content type and encoding when writing an inscription

+++ programs/libreplex_metadata/src/state/metadata.rs
@@ -69,7 +69,7 @@ pub enum Asset {
             use the chunks to calculate the PDAs off the metadata and then
             fetch all in fell swoop
         */  
-        chunks: usize,
+        chunks: u32,
     },
 }
